adding auth before getting reconnectId
